# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * vConic Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: string;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus];

export const DeviceStatus = {
  active: "active",
  inactive: "inactive",
  error: "error",
} as const;

export interface Device {
  id: string;
  name: string;
  deviceType: string;
  macAddress?: string;
  description?: string;
  token: string;
  ingestUrl: string;
  status: DeviceStatus;
  vconCount: number;
  lastSeenAt?: string;
  createdAt: string;
}

export interface DeviceList {
  devices: Device[];
  total: number;
}

export interface DeviceToken {
  token: string;
  ingestUrl: string;
}

export interface CreateDeviceRequest {
  name: string;
  deviceType: string;
  macAddress?: string;
  description?: string;
}

export type UpdateDeviceRequestStatus =
  (typeof UpdateDeviceRequestStatus)[keyof typeof UpdateDeviceRequestStatus];

export const UpdateDeviceRequestStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface UpdateDeviceRequest {
  name?: string;
  description?: string;
  status?: UpdateDeviceRequestStatus;
}

export type VconSummaryRepostStatus =
  (typeof VconSummaryRepostStatus)[keyof typeof VconSummaryRepostStatus];

export const VconSummaryRepostStatus = {
  pending: "pending",
  sent: "sent",
  failed: "failed",
  skipped: "skipped",
} as const;

export interface VconSummary {
  id: string;
  uuid: string;
  deviceId: string;
  deviceName?: string;
  duration?: number;
  partyCount: number;
  hasAnalysis: boolean;
  hasAttachments: boolean;
  repostStatus: VconSummaryRepostStatus;
  createdAt: string;
}

export interface VconList {
  vcons: VconSummary[];
  total: number;
}

export type VconPartyMeta = { [key: string]: unknown };

export interface VconParty {
  name?: string;
  tel?: string;
  mailto?: string;
  role?: string;
  meta?: VconPartyMeta;
}

export interface VconDialog {
  type: string;
  start?: string;
  duration?: number;
  parties?: number[];
  mediatype?: string;
  encoding?: string;
  body?: string;
  url?: string;
}

export interface VconAnalysis {
  type: string;
  vendor: string;
  product?: string;
  schema?: string;
  body?: string;
  encoding?: string;
  dialog?: number;
}

export interface VconAttachment {
  purpose: string;
  party?: number;
  dialog?: number;
  body?: string;
  encoding?: string;
  mediatype?: string;
  start?: string;
}

export type VconDetailRepostStatus =
  (typeof VconDetailRepostStatus)[keyof typeof VconDetailRepostStatus];

export const VconDetailRepostStatus = {
  pending: "pending",
  sent: "sent",
  failed: "failed",
  skipped: "skipped",
} as const;

export interface VconDetail {
  id: string;
  uuid: string;
  deviceId: string;
  deviceName?: string;
  vconVersion?: string;
  subject?: string;
  parties: VconParty[];
  dialog: VconDialog[];
  analysis: VconAnalysis[];
  attachments: VconAttachment[];
  repostStatus: VconDetailRepostStatus;
  repostAttempts: number;
  rawJson: string;
  createdAt: string;
}

export type VconPayloadPartiesItem = { [key: string]: unknown };

export type VconPayloadDialogItem = { [key: string]: unknown };

export type VconPayloadAnalysisItem = { [key: string]: unknown };

export type VconPayloadAttachmentsItem = { [key: string]: unknown };

export interface VconPayload {
  vcon?: string;
  uuid: string;
  created_at: string;
  parties?: VconPayloadPartiesItem[];
  dialog?: VconPayloadDialogItem[];
  analysis?: VconPayloadAnalysisItem[];
  attachments?: VconPayloadAttachmentsItem[];
  extensions?: string[];
}

export interface IngestResponse {
  id: string;
  uuid: string;
  accepted: boolean;
  message?: string;
}

export type RuleMethod = (typeof RuleMethod)[keyof typeof RuleMethod];

export const RuleMethod = {
  POST: "POST",
  PUT: "PUT",
} as const;

export type RuleHeaders = { [key: string]: string };

export interface Rule {
  id: string;
  name: string;
  deviceId?: string;
  deviceName?: string;
  targetUrl: string;
  method: RuleMethod;
  headers?: RuleHeaders;
  filterCondition?: string;
  enabled: boolean;
  successCount: number;
  failureCount: number;
  createdAt: string;
}

export interface RuleList {
  rules: Rule[];
  total: number;
}

export type CreateRuleRequestMethod =
  (typeof CreateRuleRequestMethod)[keyof typeof CreateRuleRequestMethod];

export const CreateRuleRequestMethod = {
  POST: "POST",
  PUT: "PUT",
} as const;

export type CreateRuleRequestHeaders = { [key: string]: string };

export interface CreateRuleRequest {
  name: string;
  deviceId?: string;
  targetUrl: string;
  method?: CreateRuleRequestMethod;
  headers?: CreateRuleRequestHeaders;
  filterCondition?: string;
  enabled?: boolean;
}

export type UpdateRuleRequestMethod =
  (typeof UpdateRuleRequestMethod)[keyof typeof UpdateRuleRequestMethod];

export const UpdateRuleRequestMethod = {
  POST: "POST",
  PUT: "PUT",
} as const;

export type UpdateRuleRequestHeaders = { [key: string]: string };

export interface UpdateRuleRequest {
  name?: string;
  targetUrl?: string;
  method?: UpdateRuleRequestMethod;
  headers?: UpdateRuleRequestHeaders;
  filterCondition?: string;
  enabled?: boolean;
}

export interface DashboardStats {
  totalDevices: number;
  activeDevices: number;
  totalVcons: number;
  vconsToday: number;
  totalDuration: number;
  pendingReposts: number;
  failedReposts: number;
  activeRules: number;
}

export type RecentActivityItemType =
  (typeof RecentActivityItemType)[keyof typeof RecentActivityItemType];

export const RecentActivityItemType = {
  vcon_received: "vcon_received",
  device_registered: "device_registered",
  rule_triggered: "rule_triggered",
  repost_failed: "repost_failed",
} as const;

export interface RecentActivityItem {
  id: string;
  type: RecentActivityItemType;
  message: string;
  deviceName?: string;
  vconId?: string;
  createdAt: string;
}

export interface RecentActivity {
  items: RecentActivityItem[];
}

export type ListDeviceVconsParams = {
  limit?: number;
  offset?: number;
};

export type ListVconsParams = {
  limit?: number;
  offset?: number;
};