# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: vConic Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List all devices for current user
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceList"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Register a new IoT device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceRequest"
      responses:
        "201":
          description: Device created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{deviceId}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get a device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateDevice
      tags: [devices]
      summary: Update a device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceRequest"
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Delete a device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /devices/{deviceId}/token:
    post:
      operationId: regenerateDeviceToken
      tags: [devices]
      summary: Regenerate device API token
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: New token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceToken"
  /devices/{deviceId}/vcons:
    get:
      operationId: listDeviceVcons
      tags: [vcons]
      summary: List vCons for a device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: vCon list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VconList"
  /ingest/{deviceToken}:
    post:
      operationId: ingestVcon
      tags: [ingest]
      summary: Ingest a vCon from an IoT device
      parameters:
        - name: deviceToken
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VconPayload"
      responses:
        "202":
          description: vCon accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
      required: [email, password, name]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, createdAt]
    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        deviceType:
          type: string
        macAddress:
          type: string
        description:
          type: string
        token:
          type: string
        ingestUrl:
          type: string
        status:
          type: string
          enum: [active, inactive, error]
        vconCount:
          type: integer
        lastSeenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, name, deviceType, token, ingestUrl, status, vconCount, createdAt]
    DeviceList:
      type: object
      properties:
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        total:
          type: integer
      required: [devices, total]
    DeviceToken:
      type: object
      properties:
        token:
          type: string
        ingestUrl:
          type: string
      required: [token, ingestUrl]
    CreateDeviceRequest:
      type: object
      properties:
        name:
          type: string
        deviceType:
          type: string
        macAddress:
          type: string
        description:
          type: string
      required: [name, deviceType]
    UpdateDeviceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive]
    VconSummary:
      type: object
      properties:
        id:
          type: string
        uuid:
          type: string
        deviceId:
          type: string
        deviceName:
          type: string
        duration:
          type: number
        partyCount:
          type: integer
        hasAnalysis:
          type: boolean
        hasAttachments:
          type: boolean
        repostStatus:
          type: string
          enum: [pending, sent, failed, skipped]
        createdAt:
          type: string
          format: date-time
      required: [id, uuid, deviceId, createdAt, partyCount, hasAnalysis, hasAttachments, repostStatus]
    VconList:
      type: object
      properties:
        vcons:
          type: array
          items:
            $ref: "#/components/schemas/VconSummary"
        total:
          type: integer
      required: [vcons, total]
...
    VconPayload:
      type: object
      properties:
        vcon:
          type: string
        uuid:
          type: string
        created_at:
          type: string
        parties:
          type: array
          items:
            type: object
            additionalProperties: true
        dialog:
          type: array
          items:
            type: object
            additionalProperties: true
        analysis:
          type: array
          items:
            type: object
            additionalProperties: true
        attachments:
          type: array
          items:
            type: object
            additionalProperties: true
        extensions:
          type: array
          items:
            type: string
      required: [uuid, created_at]
    IngestResponse:
      type: object
      properties:
        id:
          type: string
        uuid:
          type: string
        accepted:
          type: boolean
        message:
          type: string
      required: [id, uuid, accepted]
...