# lib/api-zod/src/generated/types/vconList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * vConic Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { VconSummary } from "./vconSummary";

export interface VconList {
  vcons: VconSummary[];
  total: number;
}