# lib/api-zod/src/generated/types/vconPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * vConic Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { VconPayloadAnalysisItem } from "./vconPayloadAnalysisItem";
import type { VconPayloadAttachmentsItem } from "./vconPayloadAttachmentsItem";
import type { VconPayloadDialogItem } from "./vconPayloadDialogItem";
import type { VconPayloadPartiesItem } from "./vconPayloadPartiesItem";

export interface VconPayload {
  vcon?: string;
  uuid: string;
  created_at: string;
  parties?: VconPayloadPartiesItem[];
  dialog?: VconPayloadDialogItem[];
  analysis?: VconPayloadAnalysisItem[];
  attachments?: VconPayloadAttachmentsItem[];
  extensions?: string[];
}